Fix: Make prodinfo check failure not an error

+++ charon/pkgs/npm.py
@@ -113,12 +113,12 @@ def handle_npm_uploading(
     succeeded = True
     for target in targets:
         manifest_folder = target[0]
-        logger.info("Start uploading manifest to s3 bucket %s", manifest_bucket_name)
         if not manifest_bucket_name:
             logger.warning(
                 'Warning: No manifest bucket is provided, will ignore the process of manifest '
                 'uploading\n')
         else:
+            logger.info("Start uploading manifest to s3 bucket %s", manifest_bucket_name)
             manifest_name, manifest_full_path = write_manifest(valid_paths, target_dir, product)
             client.upload_manifest(
                 manifest_name, manifest_full_path,
@@ -216,13 +216,18 @@ def handle_npm_del(
         )
         logger.info("Files deletion done\n")
 
-        manifest_folder = target[0]
-        logger.info(
-            "Start deleting manifest from s3 bucket %s",
-            manifest_bucket_name
-        )
-        client.delete_manifest(product, manifest_folder, manifest_bucket_name)
-        logger.info("Manifest deletion is done\n")
+        if manifest_bucket_name:
+            manifest_folder = target[0]
+            logger.info(
+                "Start deleting manifest from s3 bucket %s in folder %s",
+                manifest_bucket_name, manifest_folder
+            )
+            client.delete_manifest(product, manifest_folder, manifest_bucket_name)
+            logger.info("Manifest deletion is done\n")
+        else:
+            logger.warning(
+                'Warning: No manifest bucket is provided, will ignore the process of manifest '
+                'deletion\n')
 
         logger.info(
             "Start generating package.json for package: %s in bucket %s",
